Moved localisation helpers to shared helpers.
