--- EssentialApp/EssentialAppTests/Helpers/FeedUIIntegrationTests+Localization.swift
@@ -1,28 +1,0 @@
-//
-//  FeedViewControllerTests+Localization.swift
-//  EssentialFeediOSTests
-//
-//  Created by Pankaj Mangotra on 08/11/21.
-//
-
-import EDNLearnMac
-import Foundation
-import XCTest
-
-extension FeedUIIntegrationTests {
-    private class DummyView: ResourceView {
-        func display(_: Any) {}
-    }
-
-    var loadError: String {
-        LoadResourcePresenter<Any, DummyView>.loadError
-    }
-
-    var feedTitle: String {
-        FeedPresenter.title
-    }
-
-    var commentsTitle: String {
-        ImageCommentsPresenter.title
-    }
-}
